refactor: separate coinjoin clustering
